Updated dbc config for the shift control

+++ firmware/RAMNV1/Core/Inc/ramn_dbc.h
@@ -34,7 +34,7 @@ typedef volatile struct
 	volatile uint16_t command_accel; 			//CANID_COMMAND_ACCEL
 	volatile uint16_t control_steer; 			//CANID_CONTROL_STEERING
 	volatile uint16_t command_steer; 			//CANID_COMMAND_STEERING
-	volatile uint16_t control_shift;			//CANID_CONTROL_SHIFT
+	volatile uint8_t  control_shift;			//CANID_CONTROL_SHIFT
 	volatile uint16_t command_shift;			//CANID_COMMAND_SHIFT
 	volatile uint16_t control_sidebrake; 		//CANID_CONTROL_SIDEBRAKE
 	volatile uint16_t command_sidebrake; 		//CANID_COMMAND_SIDEBRAKE
